requirements.txt
adal==1.2.2
asn1crypto==1.3.0
astroid==2.3.3
azure-common==1.1.24
azure-core==1.2.2
azure-functions==1.2.0
azure-keyvault==4.0.0
azure-keyvault-keys==4.0.1
azure-keyvault-secrets==4.0.1
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.10.50
botocore==1.13.50
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
future==0.18.2
idna==2.8
ijson==2.6.1
isodate==0.6.0
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.11
msrestazure==0.6.2
numpy==1.18.1
oauthlib==3.1.0
oscrypto==1.2.0
pandas==0.25.3
pyarrow==0.15.1
pycparser==2.20
pycryptodomex==3.9.7
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.14.0
snowflake-connector-python==2.1.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
